Extend Decidim Initiatives Controller
